pyproject.toml
[tool.poetry]
name = "mode_wle_signer"
version = "0.1.0"
description = "Signs Mode WLE URLs"
authors = ["Irving Popovetsky <irving@honeycomb.io>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^0.21.0"
flask = "^2.3.2"
honeycomb-opentelemetry = "^0.2.1b0"
opentelemetry-instrumentation-flask = "0.38b0"
opentelemetry-instrumentation-urllib = "0.38b0"
opentelemetry-instrumentation-wsgi = "0.38b0"
opentelemetry-api = "1.17.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
